Test.Consensus.Ledger.Mock.Generators

Orphan instances

Arbitrary SimpleBody

Instance details

Methods

arbitrary ∷ Gen SimpleBody

shrink ∷ SimpleBody → [SimpleBody]

Arbitrary Tx

Instance details

Methods

arbitrary ∷ Gen Tx

shrink ∷ Tx → [Tx]

Arbitrary Addr

Instance details

Methods

arbitrary ∷ Gen Addr

shrink ∷ Addr → [Addr]

HashAlgorithm (SimpleHash c) ⇒ Arbitrary (AnnTip (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (AnnTip (SimpleBlock c ext))

shrink ∷ AnnTip (SimpleBlock c ext) → [AnnTip (SimpleBlock c ext)]

Arbitrary (GenTx (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (GenTx (SimpleBlock c ext))

shrink ∷ GenTx (SimpleBlock c ext) → [GenTx (SimpleBlock c ext)]

Arbitrary (TxId (GenTx (SimpleBlock c ext)))

Instance details

Methods

arbitrary ∷ Gen (TxId (GenTx (SimpleBlock c ext)))

shrink ∷ TxId (GenTx (SimpleBlock c ext)) → [TxId (GenTx (SimpleBlock c ext))]

Arbitrary (LedgerState (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (LedgerState (SimpleBlock c ext))

shrink ∷ LedgerState (SimpleBlock c ext) → [LedgerState (SimpleBlock c ext)]

(SimpleCrypto c, Arbitrary ext, Serialise ext, Typeable ext) ⇒ Arbitrary (Header (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (Header (SimpleBlock c ext))

shrink ∷ Header (SimpleBlock c ext) → [Header (SimpleBlock c ext)]

Arbitrary (HeaderHash blk) ⇒ Arbitrary (Point blk)

Instance details

Methods

arbitrary ∷ Gen (Point blk)

shrink ∷ Point blk → [Point blk]

Arbitrary (HeaderHash blk) ⇒ Arbitrary (ChainHash blk)

Instance details

Methods

arbitrary ∷ Gen (ChainHash blk)

shrink ∷ ChainHash blk → [ChainHash blk]

Arbitrary (MockState blk)

Instance details

Methods

arbitrary ∷ Gen (MockState blk)

shrink ∷ MockState blk → [MockState blk]

Arbitrary (HeaderHash blk) ⇒ Arbitrary (MockError blk)

Instance details

Methods

arbitrary ∷ Gen (MockError blk)

shrink ∷ MockError blk → [MockError blk]

(SimpleCrypto c, Arbitrary ext, Serialise ext) ⇒ Arbitrary (Coherent (SimpleBlock c ext))
This blindly creates random values, so the block will not be valid, but this does not matter for serialisation tests. Except we do touch-up the simpleBodySize; hence Coherent.

Instance details

Methods

arbitrary ∷ Gen (Coherent (SimpleBlock c ext))

shrink ∷ Coherent (SimpleBlock c ext) → [Coherent (SimpleBlock c ext)]

(SimpleCrypto c, Typeable ext) ⇒ Arbitrary (SomeResult (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (SomeResult (SimpleBlock c ext))

shrink ∷ SomeResult (SimpleBlock c ext) → [SomeResult (SimpleBlock c ext)]

Arbitrary (BftFields BftMockCrypto toSign)

Instance details

Methods

arbitrary ∷ Gen (BftFields BftMockCrypto toSign)

shrink ∷ BftFields BftMockCrypto toSign → [BftFields BftMockCrypto toSign]

Arbitrary (SomeSecond BlockQuery (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (SomeSecond BlockQuery (SimpleBlock c ext))

shrink ∷ SomeSecond BlockQuery (SimpleBlock c ext) → [SomeSecond BlockQuery (SimpleBlock c ext)]

Arbitrary (SomeSecond (NestedCtxt Header) (SimpleBlock c ext))

Instance details

Methods

arbitrary ∷ Gen (SomeSecond (NestedCtxt Header) (SimpleBlock c ext))

shrink ∷ SomeSecond (NestedCtxt Header) (SimpleBlock c ext) → [SomeSecond (NestedCtxt Header) (SimpleBlock c ext)]

Arbitrary (SimpleBftExt c BftMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (SimpleBftExt c BftMockCrypto)

shrink ∷ SimpleBftExt c BftMockCrypto → [SimpleBftExt c BftMockCrypto]

(SimpleCrypto c, Arbitrary ext, Serialise ext) ⇒ Arbitrary (SimpleBlock c ext)
This blindly creates random values, so the block will not be valid, but this does not matter for serialisation tests.

Instance details

Methods

arbitrary ∷ Gen (SimpleBlock c ext)

shrink ∷ SimpleBlock c ext → [SimpleBlock c ext]

(HashAlgorithm (SimpleHash c), Arbitrary ext, Serialise ext) ⇒ Arbitrary (SimpleStdHeader c ext)

Instance details

Methods

arbitrary ∷ Gen (SimpleStdHeader c ext)

shrink ∷ SimpleStdHeader c ext → [SimpleStdHeader c ext]

HashAlgorithm h ⇒ Arbitrary (Hash h a)

Instance details

Methods

arbitrary ∷ Gen (Hash h a)

shrink ∷ Hash h a → [Hash h a]